requirements.txt
django
watchdog
gitpython
langchain
openai
chromadb
python-dotenv
pydantic